# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kenaz-ml",
    "description": "kenaz-ml \u2014 the ML sidecar for Sigil (local mode)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/predict/stuck": {
      "post": {
        "summary": "Predict Stuck",
        "operationId": "predict_stuck_predict_stuck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StuckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StuckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/suggest": {
      "post": {
        "summary": "Predict Suggest",
        "operationId": "predict_suggest_predict_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/duration": {
      "post": {
        "summary": "Predict Duration",
        "operationId": "predict_duration_predict_duration_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/quality": {
      "post": {
        "summary": "Predict Quality",
        "operationId": "predict_quality_predict_quality_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/train": {
      "post": {
        "summary": "Train",
        "operationId": "train_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DurationRequest": {
        "properties": {
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "features": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          }
        },
        "type": "object",
        "title": "DurationRequest"
      },
      "DurationResponse": {
        "properties": {
          "estimated_minutes": {
            "type": "number",
            "title": "Estimated Minutes"
          },
          "confidence_interval": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Confidence Interval"
          }
        },
        "type": "object",
        "required": [
          "estimated_minutes",
          "confidence_interval"
        ],
        "title": "DurationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QualityRequest": {
        "properties": {
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "QualityRequest"
      },
      "QualityResponse": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Components"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "score",
          "components",
          "status"
        ],
        "title": "QualityResponse"
      },
      "StuckRequest": {
        "properties": {
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "features": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          }
        },
        "type": "object",
        "title": "StuckRequest"
      },
      "StuckResponse": {
        "properties": {
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "probability",
          "confidence"
        ],
        "title": "StuckResponse"
      },
      "TrainRequest": {
        "properties": {
          "db": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Db",
            "description": "Deprecated: ignored, kept for backward compat"
          }
        },
        "type": "object",
        "title": "TrainRequest"
      },
      "TrainResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "TrainResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowStateRequest": {
        "properties": {
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "classified_events": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classified Events"
          }
        },
        "type": "object",
        "title": "WorkflowStateRequest"
      },
      "WorkflowStateResponse": {
        "properties": {
          "flow_state": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Flow State"
          },
          "dominant_state": {
            "type": "string",
            "title": "Dominant State"
          },
          "momentum": {
            "type": "number",
            "title": "Momentum"
          },
          "focus_score": {
            "type": "number",
            "title": "Focus Score"
          },
          "dominant_activity": {
            "type": "string",
            "title": "Dominant Activity"
          },
          "activity_distribution": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Activity Distribution"
          },
          "session_elapsed_min": {
            "type": "number",
            "title": "Session Elapsed Min"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "flow_state",
          "dominant_state",
          "momentum",
          "focus_score",
          "dominant_activity",
          "activity_distribution",
          "session_elapsed_min",
          "method",
          "confidence"
        ],
        "title": "WorkflowStateResponse"
      }
    }
  }
}